overlay: Extract count overlay logic from `IconUnreadConversations`
Create new component `CountOverlay`, which can be used in case we want
count overlay on any component. It internally uses `ComponentWithOverlay`
with specific props.
