+++ src/common/index.js
@@ -5,6 +5,7 @@ export { default as Centerer } from './Centerer';
 export { default as Circle } from './Circle';
 export { default as ComponentList } from './ComponentList';
 export { default as ComponentWithOverlay } from './ComponentWithOverlay';
+export { default as CountOverlay } from './CountOverlay';
 export { default as ErrorMsg } from './ErrorMsg';
 export { default as FloatingActionButton } from './FloatingActionButton';
 export { default as UserAvatar } from './UserAvatar';
