+++ src/nav/IconUnreadConversations.js
@@ -1,20 +1,13 @@
 /* @flow strict-local */
 
 import React, { PureComponent } from 'react';
-import { StyleSheet, View } from 'react-native';
+import { View } from 'react-native';
 
 import type { Dispatch } from '../types';
 import { connect } from '../react-redux';
 import { getUnreadPmsTotal, getUnreadHuddlesTotal } from '../selectors';
 import { IconPeople } from '../common/Icons';
-import { ComponentWithOverlay, UnreadCount } from '../common';
-import { BRAND_COLOR } from '../styles';
-
-const styles = StyleSheet.create({
-  button: {
-    flex: 1,
-  },
-});
+import { CountOverlay } from '../common';
 
 type Props = {|
   dispatch: Dispatch,
@@ -30,16 +23,9 @@ class IconUnreadConversations extends PureComponent<Props> {
 
     return (
       <View>
-        <ComponentWithOverlay
-          style={styles.button}
-          overlaySize={15}
-          overlayColor={BRAND_COLOR}
-          overlayPosition="top-right"
-          showOverlay={unreadCount > 0}
-          overlay={<UnreadCount count={unreadCount} />}
-        >
+        <CountOverlay unreadCount={unreadCount}>
           <IconPeople size={24} color={color} />
-        </ComponentWithOverlay>
+        </CountOverlay>
       </View>
     );
   }
